refactor: Extract initApp into its own files
